requirements.txt
Flask[async]==3.*
SQLAlchemy==2.*
Flask-SQLAlchemy==3.*
Alembic==1.*
Flask-Migrate==4.*
Flask-Login==0.6.*
Flask-WTF==1.*
email-validator==2.*
Flask-Limiter==3.*
python-dotenv==1.*
bcrypt==4.*
gunicorn==21.2.0
requests==2.32.2
psycopg2-binary>=2.9.9
pytest==8.*
redis==5.*
rq==1.*

